refactor: Change response package path
