+++ api/src/main/java/grooteogi/controller/HashtagController.java
@@ -2,7 +2,7 @@ package grooteogi.controller;
 
 import grooteogi.domain.Hashtag;
 import grooteogi.dto.HashtagDto;
-import grooteogi.dto.response.BasicResponse;
+import grooteogi.response.BasicResponse;
 import grooteogi.service.HashtagService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
